Fixed range_decoder for web
